requirements.txt
matplotlib==3.4.2
numpy==1.20.3
pandas==1.3.1
pandas-datareader==0.10.0
pyyaml